major re-org
Moving to a layered system design.

Server
- http
- chsk

Communications
- gateway
- private

Data
- core
- datomic

The server layer interfaces between the client applications and the
server-side system. This is a restful abstraction over the machinery
for communications with client applications.

The communications layer interfaces between the server endpoints and
the rest of the system. This is a set of semantic async pathways
connecting the many system components.

The data layer interfaces between the database and the server-side
system. This is an abstraction over the underlying data model and
storage implementation.